feat(console): add validation detail, reorg, tests
