+++ src/internal/tui/component/focus.go
@@ -1,0 +1,15 @@
+package component
+
+type focus int
+
+const (
+	noComponentFocus focus = iota
+	focusComponentSelection
+	focusFrameworkSelection
+	focusControls
+	focusRemarks
+	focusDescription
+	focusValidations
+)
+
+var maxFocus = focusValidations
